refactor(language): extract template subsystem into dedicated template.ts
Moved TemplateText, TemplateInterpolation, TemplateExpression, parseTemplateParts,
TEMPLATE_PART_KINDS, isTemplatePartKind, and all dedent/clean helpers
(~390 lines) out of expressions.ts into a new core/template.ts. The
template subsystem had zero coupling to other expression classes — it only
used the Expression type for its interpolation payload and extended
ExpressionBase — so the split is clean.

Also extracted the Expression interface and ExpressionBase abstract class
into expression-base.ts to avoid the runtime import cycle that would
otherwise result (template.ts needs ExpressionBase; expressions.ts
re-exports TemplateExpression). expression-base.ts is a leaf file imported
by both expressions.ts and template.ts.

expressions.ts re-exports the template primitives so existing consumers
keep working without import-path changes; public surface is unchanged.
